take shaders out of main.ts
