+++ README.md
@@ -1,6 +1,8 @@
 # KSU Hackathon 2025
 
-To start server on http://localhost:3000 
+Start server on http://localhost:3000 
 ```
 npm run dev
 ```
+
+code from https://webgl2fundamentals.org/webgl/lessons/webgl-fundamentals.html
