+++ src/main.ts
@@ -1,61 +1,8 @@
-function createShader(
-  gl: WebGL2RenderingContext,
-  type: number,
-  source: string
-) {
-  const shader = gl.createShader(type);
-  if (shader == null) {
-    throw new Error("GL shader was null.");
-  }
+import { resizeCanvasToDisplaySize, createShader, createProgram } from './util';
 
-  gl.shaderSource(shader, source);
-  gl.compileShader(shader);
-
-  const success = gl.getShaderParameter(shader, gl.COMPILE_STATUS)
-  if (success) {
-    return shader
-  }
-
-  console.log(gl.getShaderInfoLog(shader));
-  gl.deleteShader(shader);
-}
-
-function createProgram(
-  gl: WebGL2RenderingContext,
-  vertexShader: WebGLShader,
-  fragmentShader: WebGLShader,
-) {
-  const program = gl.createProgram();
-  gl.attachShader(program, vertexShader);
-  gl.attachShader(program, fragmentShader);
-  gl.linkProgram(program);
-
-  const success = gl.getProgramParameter(program, gl.LINK_STATUS);
-  if (success) {
-    return program;
-  }
-
-  console.log(gl.getProgramInfoLog(program));
-  gl.deleteProgram(program);
-}
-
-function resizeCanvasToDisplaySize(canvas: HTMLCanvasElement | null): boolean {
-  if (!canvas) {
-    return false;
-  }
-
-  const displayWidth = canvas.clientWidth;
-  const displayHeight = canvas.clientHeight;
-
-  const needResize = canvas.width !== displayWidth || canvas.height !== displayHeight;
-
-  if (needResize) {
-    canvas.width = displayWidth;
-    canvas.height = displayHeight;
-  }
-
-  return needResize;
-}
+// allowed by the flags provided to esbuild, ignore err
+import vertexShaderSource from './shaders/shader.vert'; // --loader:.vert=text flag 
+import fragmentShaderSource from './shaders/shader.frag'; // --loader:.frag=text flag 
 
 function main() {
   const canvas = document.getElementById("canvas") as HTMLCanvasElement | null;
@@ -64,57 +11,18 @@ function main() {
   const gl = canvas.getContext("webgl2");
   if (gl == null) throw new Error("Couldn't get webgl2 context from canvas.");
 
-  const vertexShaderSource = `#version 300 es
-    
-  // an attribute is an input (in) to a vertex shader.
-  // It will receive data from a buffer
-  in vec4 a_position;
-    
-  // all shaders have a main function
-  void main() {
-    
-    // gl_Position is a special variable a vertex shader
-    // is responsible for setting
-    gl_Position = a_position;
-  }
-  `;
-
-  const fragmentShaderSource = `#version 300 es
-    
-  // fragment shaders don't have a default precision so we need
-  // to pick one. highp is a good default. It means "high precision"
-  precision highp float;
-    
-  // we need to declare an output for the fragment shader
-  out vec4 outColor;
-    
-  void main() {
-    // Just set the output to a constant reddish-purple
-    outColor = vec4(1, 0, 0.5, 1);
-  }
-  `;
-
   const vertexShader = createShader(gl, gl.VERTEX_SHADER, vertexShaderSource);
   const fragmentShader = createShader(gl, gl.FRAGMENT_SHADER, fragmentShaderSource);
 
-  if (vertexShader == undefined) {
-    throw new Error("Vertex shader is undefined.");
-  }
-
-  if (fragmentShader == undefined) {
-    throw new Error("Fragment shader is undefined.");
-  }
+  if (vertexShader == undefined) throw new Error("Vertex shader is undefined.");
+  if (fragmentShader == undefined) throw new Error("Fragment shader is undefined.");
 
   var program = createProgram(gl, vertexShader, fragmentShader);
 
-  if (program == undefined) {
-    throw new Error("WebGL2 program is undefined.");
-  }
+  if (program == undefined) throw new Error("WebGL2 program is undefined.");
 
   const positionAttributeLocation = gl.getAttribLocation(program, "a_position");
-
   const positionBuffer = gl.createBuffer();
-
   gl.bindBuffer(gl.ARRAY_BUFFER, positionBuffer);
 
   const positions = [
@@ -137,7 +45,7 @@ function main() {
     positionAttributeLocation, size, type, normalize, stride, attrib_offset
   )
 
-  resizeCanvasToDisplaySize(gl.canvas as HTMLCanvasElement | null);
+  resizeCanvasToDisplaySize(gl.canvas as HTMLCanvasElement | null)
 
   gl.viewport(0, 0, gl.canvas.width, gl.canvas.height);
   gl.clearColor(0, 0, 0, 0);
