+++ src/util.ts
@@ -1,0 +1,58 @@
+export function resizeCanvasToDisplaySize(canvas: HTMLCanvasElement | null): boolean {
+  if (canvas == null) {
+    return false;
+  }
+
+  const displayWidth = canvas.clientWidth;
+  const displayHeight = canvas.clientHeight;
+
+  const needResize = canvas.width !== displayWidth || canvas.height !== displayHeight;
+
+  if (needResize) {
+    canvas.width = displayWidth;
+    canvas.height = displayHeight;
+  }
+
+  return needResize;
+}
+
+export function createShader(
+  gl: WebGL2RenderingContext,
+  type: number,
+  source: string
+) {
+  const shader = gl.createShader(type);
+  if (shader == null) {
+    throw new Error("GL shader was null.");
+  }
+
+  gl.shaderSource(shader, source);
+  gl.compileShader(shader);
+
+  const success = gl.getShaderParameter(shader, gl.COMPILE_STATUS)
+  if (success) {
+    return shader
+  }
+
+  console.log(gl.getShaderInfoLog(shader));
+  gl.deleteShader(shader);
+}
+
+export function createProgram(
+  gl: WebGL2RenderingContext,
+  vertexShader: WebGLShader,
+  fragmentShader: WebGLShader,
+) {
+  const program = gl.createProgram();
+  gl.attachShader(program, vertexShader);
+  gl.attachShader(program, fragmentShader);
+  gl.linkProgram(program);
+
+  const success = gl.getProgramParameter(program, gl.LINK_STATUS);
+  if (success) {
+    return program;
+  }
+
+  console.log(gl.getProgramInfoLog(program));
+  gl.deleteProgram(program);
+}
